engine update: -added console command feature -added button onClick handler -few minor changes -no need to tick/draw console in game
git-svn-id: https://happy-engine.googlecode.com/svn/trunk@163 5aa1a52c-6ec2-ce28-c747-3d7cbe9a755d

+++ src/HappyEngine/code/graphics/RoundedRectangle2D.h
@@ -36,7 +36,7 @@ public:
 	/* CONSTRUCTOR - DESTRUCTOR */
 	RoundedRectangle2D(	const vec2& centerPos,
 						const vec2& size,
-						uint radius,
+						float radius,
 						uint interpolSteps = 30);
     virtual ~RoundedRectangle2D();
 };
